folder name is updated
