+++ SecondYear/29deepanshutyagi_Deepanshu Tyagi_2226cs1195_2/week_1/pattern 10.cpp
@@ -19,16 +19,16 @@ public:
 
 //{ Driver Code Starts.
 
-int main() {
-    int t;
-    cin >> t;
-    while (t--) {
-        int n;
-        cin >> n;
+// int main() {
+//     int t;
+//     cin >> t;
+//     while (t--) {
+//         int n;
+//         cin >> n;
         
-        Solution ob;
-        ob.printTriangle(n);
-    }
-    return 0;
-}
+//         Solution ob;
+//         ob.printTriangle(n);
+//     }
+//     return 0;
+// }
 // } Driver Code Ends
