+++ SecondYear/29deepanshutyagi_Deepanshu Tyagi_2226cs1195_2/week_1/pattern 7.cpp
@@ -48,16 +48,16 @@ class Solution {
 
 //{ Driver Code Starts.
 
-int main() {
-    int t;
-    cin >> t;
-    while (t--) {
-        int n;
-        cin >> n;
+// int main() {
+//     int t;
+//     cin >> t;
+//     while (t--) {
+//         int n;
+//         cin >> n;
 
-        Solution ob;
-        ob.printDiamond(n);
-    }
-    return 0;
-}
-// } Driver Code Ends
+//         Solution ob;
+//         ob.printDiamond(n);
+//     }
+//     return 0;
+// }
+// // } Driver Code Ends
