+++ SecondYear/29deepanshutyagi_Deepanshu Tyagi_2226cs1195_2/week_1/pattern 9.cpp
@@ -50,16 +50,16 @@ class Solution {
 
 //{ Driver Code Starts.
 
-int main() {
-    int t;
-    cin >> t;
-    while (t--) {
-        int n;
-        cin >> n;
+// int main() {
+//     int t;
+//     cin >> t;
+//     while (t--) {
+//         int n;
+//         cin >> n;
 
-        Solution ob;
-        ob.printTriangle(n);
-    }
-    return 0;
-}
+//         Solution ob;
+//         ob.printTriangle(n);
+//     }
+//     return 0;
+// }
 // } Driver Code Ends
